+++ pages/auth/_token.vue
@@ -3,11 +3,18 @@
 </template>
 
 <script>
+import getUserInfo from '@/apollo/queries/getUserInfo.gql'
+
 export default {
   mounted() {
     const token = this.$route.params.token
     if (token) this.$apolloHelpers.onLogin(token)
-    this.$router.replace('/')
+    this.$router.replace(`/user/${this.me.login}`)
+  },
+  apollo: {
+    me: {
+      query: getUserInfo
+    }
   }
 }
 </script>